Add files via upload
Biomedical ontology is getting important. It integrates the information about various model organisms, obtains under different conditions and stores in heterogeneous database. A controlled vocabulary of biomedical terms (which are organized in order to allow the expression of complex relationships in machine-readable form) is needed for biomedical ontology. Biomarkers have demonstrated effectiveness in quantifying biological molecules that can be found, which not only reveal normal or abnormal processes but also a condition of cancer.